Refactoring:
  - `PluginsIcons` migrated to kotlin
  -  Refactored affected files.
