--- src/main/kotlin/no/tornado/tornadofx/idea/PluginIcons.java
@@ -1,9 +1,0 @@
-package no.tornado.tornadofx.idea;
-
-import com.intellij.openapi.util.IconLoader;
-
-import javax.swing.*;
-
-public interface PluginIcons {
-	Icon ACTION = IconLoader.getIcon("/icons/action.png");
-}
