+++ src/main/kotlin/no/tornado/tornadofx/idea/framework/TornadoFXFrameworkSupportProvider.kt
@@ -3,7 +3,7 @@ package no.tornado.tornadofx.idea.framework
 import com.intellij.ide.util.frameworkSupport.FrameworkSupportModel
 import com.intellij.ide.util.frameworkSupport.FrameworkSupportProvider
 import com.intellij.openapi.module.ModuleType
-import no.tornado.tornadofx.idea.PluginIcons
+import no.tornado.tornadofx.idea.icons.PluginIcons
 import no.tornado.tornadofx.idea.TornadoFXModuleType
 import no.tornado.tornadofx.idea.facet.TornadoFXFacetType
 import javax.swing.Icon
