+++ src/main/kotlin/no/tornado/tornadofx/idea/icons/PluginIcons.kt
@@ -1,0 +1,7 @@
+package no.tornado.tornadofx.idea.icons
+
+import com.intellij.openapi.util.IconLoader
+
+object PluginIcons {
+    val ACTION = IconLoader.getIcon("/icons/action.png");
+}
